+++ src/texture/t2d.rs
@@ -10,6 +10,7 @@ use crate::texture::inner::TextureAccessible;
 use crate::Extent2d;
 
 /// 2d texture
+#[repr(transparent)]
 pub struct Texture2D {
     ffi: gli::texture2d,
 }